refactor: Modify grpc fs to make adding more services easier
