+++ runner/src/index.ts
@@ -1,10 +1,10 @@
 import { startServer as startMetricsServer } from './metrics';
-import startRunnerServer from './server/runner-server';
+import { startServer as startGrpcServer } from './server';
 import type StreamHandler from './stream-handler';
 import logger from './logger';
 
 const executors = new Map<string, StreamHandler>();
-startRunnerServer(executors);
+startGrpcServer(executors);
 
 startMetricsServer().catch((err) => {
   logger.error('Failed to start metrics server', err);
