+++ runner/src/server/services/runner/index.ts
@@ -1,0 +1,1 @@
+export { getRunnerService } from './runner-service';
