+++ runner/src/server/services/runner/runner-client.ts
@@ -1,7 +1,7 @@
 import * as grpc from '@grpc/grpc-js';
 import * as protoLoader from '@grpc/proto-loader';
-import { type ProtoGrpcType } from '../generated/runner';
-import { type RunnerClient } from '../generated/runner/Runner';
+import { type ProtoGrpcType } from '../../../generated/runner';
+import { type RunnerClient } from '../../../generated/runner/Runner';
 
 const PROTO_PATH = 'protos/runner.proto';
